# docs/swagger.yaml
basePath: /
definitions:
  domain.Category:
    properties:
      category_name:
        type: string
      id:
        type: integer
    required:
    - category_name
    type: object
  domain.Products:
    properties:
      brand_id:
        type: integer
      category_id:
        type: integer
      design_description:
        type: string
      id:
        type: integer
      is_deleted:
        type: boolean
      name:
        type: string
      price:
        type: number
      product_status:
        type: string
      quantity:
        type: integer
      sku:
        type: string
    required:
    - name
    type: object
...
  models.ProductUpdate:
    properties:
      product_id:
        type: integer
      quantity:
        type: integer
    type: object
  models.SetNewName:
    properties:
      current:
        type: string
      new:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
host: www.zogfestiv.store
info:
  contact: {}
  description: API for ecommerce website
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Zog_festiv eCommerce API
  version: "1.0"
paths:
...
  /admin/category:
    post:
      consumes:
      - application/json
      description: Add a new Category for more contents
      parameters:
      - description: add Category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/domain.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Add a new Category
      tags:
      - Admin Category Management
    put:
      consumes:
      - application/json
      description: updating the current Catergory
      parameters:
      - description: update Category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.SetNewName'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: update a  Category
      tags:
      - Admin Category Management
  /admin/category/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Category for existing products that belongs to the Category
      parameters:
      - description: category-id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Delete Category
      tags:
      - Admin Category Management
...
  /admin/products:
    post:
      consumes:
      - application/json
      description: Add product from admin side
      parameters:
      - description: Product details
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/domain.Products'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Add Products
      tags:
      - Admin Product Management
    put:
      consumes:
      - application/json
      description: Update quantity of already existing product
      parameters:
      - description: Product details
        in: body
        name: productUpdate
        required: true
        schema:
          $ref: '#/definitions/models.ProductUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Update Products quantity
      tags:
      - Admin Product Management
  /admin/products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product from the admin side
      parameters:
      - description: product id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Delete product
      tags:
      - Admin Product Management
...
  /user/products/{id}:
    get:
      consumes:
      - application/json
      description: Get Individual Detailed product details to user side
      parameters:
      - description: idstr
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get Individual Product Details
      tags:
      - User Product
  /user/products/filter:
    post:
      consumes:
      - application/json
      description: Show all the Products belonging to a specified category
      parameters:
      - description: Category IDs and quantities
        in: body
        name: data
        required: true
        schema:
          additionalProperties:
            type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Show Products of specified category
      tags:
      - User Product
  /user/products/page/{page}:
    get:
      consumes:
      - application/json
      description: Retrieve all product Details with pagination to users
      parameters:
      - description: Page number
        in: path
        name: page
        required: true
        type: string
      - description: Page Count
        in: query
        name: count
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get Products Details to users
      tags:
      - User Product
...
schemes:
- http
securityDefinitions:
  JWT:
    in: header
    name: token
    type: apiKey
swagger: "2.0"